requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.0,<4.0.0

# Code Quality
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.12.0,<6.0.0
mypy>=1.5.0,<2.0.0
pylint>=2.17.0,<3.0.0

# Pre-commit hooks
pre-commit>=3.3.0,<4.0.0

# Type stubs
types-PyYAML>=6.0.0,<7.0.0
pandas-stubs>=2.0.0,<3.0.0
